requirements.txt
django-tastypie==0.14.4